fixed to BigDecimal

+++ src/main/java/com/malikov/shopsystem/service/OrderItemService.java
@@ -19,8 +19,8 @@ public interface OrderItemService{
 
     void updateProductName(Long id, String newProductName);
 
-    int updateOrderItemProductQuantity(Long itemId, int quantity);
-    int updateOrderItemProductPrice(Long itemId, BigDecimal price);
+    BigDecimal updateOrderItemProductQuantity(Long itemId, int quantity);
+    BigDecimal updateOrderItemProductPrice(Long itemId, BigDecimal price);
 
     OrderItem createNewEmpty(Long orderId);
